% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEObridge.R
\name{barplot_interact}
\alias{barplot_interact}
\title{The interactive plot function for plotting the interested platform of user selected}
\usage{
barplot_interact(top10platform)
}
\arguments{
\item{top10platform}{A variable returned from function of print_top_N_platform.}
}
\description{
The interactive plot function for plotting the interested platform of user selected
}
