% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEObridge.R
\name{barplot_top_N_platform}
\alias{barplot_top_N_platform}
\title{The function to show the barplot of the most N frequent platforms in current dataset.}
\usage{
barplot_top_N_platform(top10platform)
}
\arguments{
\item{top10platform}{A variable returned from function of print_top_N_platform.}
}
\value{
A barplot figure.
}
\description{
The function to show the barplot of the most N frequent platforms in current dataset.
}
