% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEObridge.R
\name{keyword_search}
\alias{keyword_search}
\title{The main function of mining the latent relationship among variables.}
\usage{
keyword_search(
  key1 = "breast",
  key2 = "cancer",
  limitN = 2000,
  verbose = TRUE,
  db = connection,
  subTable = "gse"
)
}
\arguments{
\item{key1}{The first keyword to search. For example we want to search "breast cancer", we can set key1 is the first word, breast.}

\item{key2}{The second keyword to search. For example we want to search "breast cancer", we can set key2 is the second word, cancer}

\item{limitN}{An integer variable to choose largest number query from the database.}

\item{verbose}{A logical variable. It will print some disgnostic information if TRUE.}

\item{db}{The connection variable which contains all metadata.}

\item{subTable}{Character variable. It is the sub-table name which stores different types of metadata.}
}
\value{
The returned result is GSE ID of the dataset which satisfy the search rules.
}
\description{
The main function of mining the latent relationship among variables.
}
