% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEObridge.R
\name{prepare_table}
\alias{prepare_table}
\title{This function has two goals. First, prepare the database locally for downstream searching.
Second, print the detailed columns in subTable (e.g. gse).}
\usage{
prepare_table(printAllTable = TRUE, subTable = "gse")
}
\arguments{
\item{printAllTable}{logical variable. Print all table if TRUE, do not print all table if FALSE.}

\item{subTable}{Character variable. It is the sub-table name which stores different types of metadata. It should be one of followings. ('gds', 'gds_subset', 'geoConvert','geodb_column_desc','gpl','gse','gse_gpl','gse_gsm','gsm','metaInfo','sMatrix')}
}
\value{
The connection variable which contains all metadata.
}
\description{
This function has two goals. First, prepare the database locally for downstream searching.
Second, print the detailed columns in subTable (e.g. gse).
}
